# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a backend server for alfarpos.",
        "title": "AlfarPOS BackEnd",
        "contact": {
            "name": "Almantera Tiantana Al Faruqi",
            "url": "https://twitter.com/almanalfaruq",
            "email": "alman.alfaruq@gmail.com"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/",
    "paths": {
        "/money": {
            "post": {
                "description": "New money is used for creating a new money transaction (in/out)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "money"
                ],
                "summary": "New money",
                "parameters": [
                    {
                        "description": "Money transaction",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/transaction.Money"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Return money transaction data",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.ResponseStruct"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/transaction.Money"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "Return error with message",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.ResponseStruct"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Return error with message",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.ResponseStruct"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/money/filters": {
            "post": {
                "description": "Get money transaction with filter status and date",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "money"
                ],
                "summary": "Get money transaction with filter",
                "parameters": [
                    {
                        "description": "Money transaction",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/transaction.GetMoneyWithFilterReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Return monies transaction data",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.ResponseStruct"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/transaction.Money"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "Return error with message",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.ResponseStruct"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Return error with message",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.ResponseStruct"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "response.ResponseStruct": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 200
                },
                "data": {},
                "message": {
                    "type": "string",
                    "example": "Success getting all products"
                }
            }
        },
...
        "transaction.GetMoneyWithFilterReq": {
            "type": "object",
            "properties": {
                "end_date": {
                    "type": "string"
                },
                "start_date": {
                    "type": "string"
                },
                "types": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
        "transaction.Money": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number",
                    "example": 1000000
                },
                "created_at": {
                    "type": "string"
                },
                "deleted_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "note": {
                    "type": "string",
                    "example": "Payment for unilever"
                },
                "type": {
                    "type": "integer",
                    "example": 1
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
    }
}